Add nf-float v0.4.3
Upgrade nf-float plugin to 0.4.3. This version includes a fix that
sometimes the job status in MMC is not updated to Nextflow properly.

The root cause is that we depend on a full job list for individual
job status update.  But sometimes the full poll doesn't take place
or doesn't finish in time.

The fix is to create the status entry for each job in the plugin
right after we submit a job to MMC.

The change is available at:
https://github.com/MemVerge/nf-float/pull/74